Fixes from tfs 0.2 *Fixed a trade bug *Fixed possible data losses *Fixed const now TEXTCOLOR_DARKRED and TEXTCOLOR_PURPLE are correct text color of COMBAT_ICEDAMAGE should be SKYBLUE . Credits goes to TFS 0.2 Team (Talaturen).

+++ otserv/trunk/combat.cpp
@@ -1437,10 +1437,10 @@ void AreaCombat::copyArea(const MatrixArea* input, MatrixArea* output, MatrixOpe
 		}
 		double angleRad = 3.1416 * angle / 180.0;
 
-		float a = std::cos(angleRad);
-		float b = -std::sin(angleRad);
-		float c = std::sin(angleRad);
-		float d = std::cos(angleRad);
+		double a = std::cos(angleRad);
+		double b = -std::sin(angleRad);
+		double c = std::sin(angleRad);
+		double d = std::cos(angleRad);
 
 		for(int32_t x = 0; x < (int32_t)input->getCols(); ++x){
 			for(int32_t y = 0; y < (int32_t)input->getRows(); ++y){
